refactor(adr-159): address code quality review
- snapshots report the position the screen applied, not the position broadcast:
  a timed-out headless flush was telling clients to drop output nothing could
  rebuild
- syncing a terminal with its session is one operation again — openRestored
  writes the snapshot and releases the queue, replacing a mutable let bridging
  .then and .finally; the queue re-arms per attach rather than once per mount
- revert the seq plumbing added to electron/pty.ts, which nothing imports
- one spelling of stream position across the boundary, dropping the cast that
  built a StreamEvent the type forbade
- extract the daemon test harness so protocol tests get their own file; e2e.test
  drops from 1292 to 881 lines
- e2e: strip VITE_DEV_SERVER_URL from the app under test and share one
  launch/teardown; wait for the shell to settle instead of retyping into it
