+++ electron/ipc/pty.ts
@@ -76,8 +76,9 @@ export function register(deps: IpcDeps): void {
           ok: true,
           snapshot: result.snapshot?.screenAnsi || null,
           // Stream position the snapshot reflects — the renderer drops queued
-          // output at or below it (ADR-159).
-          snapshotSeq: result.snapshot?.seq ?? null,
+          // output at or below it (ADR-159). Absent when there is no snapshot,
+          // or when an older daemon does not report one.
+          snapshotSeq: result.snapshot?.seq,
           prewarmed: result.snapshot !== null,
         };
       } catch (err) {
