+++ src/hooks/useTerminalStream.ts
@@ -1,12 +1,15 @@
 /**
  * useTerminalStream — subscribes to PTY output, exit, and CWD events.
  *
- * Owns the whole PTY-bytes-to-xterm path, including its opening move: output is
- * queued until the caller calls `openOutput(term, snapshotSeq)`. The daemon
+ * Owns the whole PTY-bytes-to-xterm path, including its opening move: nothing
+ * reaches the terminal until `openRestored()` puts it in sync with the daemon.
+ *
+ * That opening move is one operation, so it lives in one place. The daemon
  * subscribes us to the session's stream before the create/attach reply gets
  * back here — deliberately, so nothing is lost in the gap — which means some of
- * what we queued is also inside the warm-restore snapshot. Each chunk's stream
- * position sorts that out: see `outputAfterSnapshot`.
+ * the output we queued is also inside the warm-restore snapshot. `openRestored`
+ * writes the snapshot and then only the queued chunks it does not already
+ * cover; see `outputAfterSnapshot`.
  *
  * Also handles kitty keyboard protocol negotiation: intercepts push/pop/query
  * sequences from the child process and responds on behalf of xterm.js (which
@@ -20,6 +23,7 @@ import { useMountEffect } from "./useMountEffect";
 import {
   outputAfterSnapshot,
   type QueuedOutput,
+  type TerminalRestore,
 } from "../lib/snapshot-dedupe";
 
 // Matches kitty keyboard protocol sequences:
@@ -45,23 +49,30 @@ export function useTerminalStream(
   const queuedRef = useRef<QueuedOutput[] | null>([]);
 
   /**
-   * Start writing output through, replaying what arrived while queued minus
-   * whatever the snapshot at `snapshotSeq` already put on screen.
+   * Bring `target` up to the session's current state and let output flow: write
+   * the warm-restore snapshot, then the queued chunks it does not already
+   * account for.
    *
    * Takes the terminal explicitly so the caller doesn't have to assume React
-   * has re-rendered with it by the time the create/attach round trip lands.
+   * has re-rendered with it by the time the create/attach round trip lands. A
+   * cold session passes no snapshot, which writes the queue as-is.
    */
-  const openOutput = useCallback(
-    (target: Terminal, snapshotSeq?: number | null) => {
-      const queued = queuedRef.current;
+  const openRestored = useCallback(
+    (target: Terminal, snapshot?: TerminalRestore | null) => {
+      const queued = queuedRef.current ?? [];
       queuedRef.current = null;
-      if (!queued) return;
-      for (const chunk of outputAfterSnapshot(queued, snapshotSeq)) {
+      if (snapshot?.ansi) target.write(snapshot.ansi);
+      for (const chunk of outputAfterSnapshot(queued, snapshot?.seq)) {
         target.write(chunk.data);
       }
     },
     [],
   );
+
+  /** Queue output again, so a re-attach gets the same ordering guarantee. */
+  const closeOutput = useCallback(() => {
+    queuedRef.current = [];
+  }, []);
 
   useMountEffect(() => {
     let kittyFlags = 0;
@@ -124,5 +135,5 @@ export function useTerminalStream(
     };
   });
 
-  return { openOutput };
+  return { openRestored, closeOutput };
 }
