+++ tests/e2e/background-setup.spec.ts
@@ -1,13 +1,13 @@
 import fs from "fs";
 import path from "path";
+import { type ElectronApplication, type Page } from "@playwright/test";
 import {
-  _electron,
-  type ElectronApplication,
-  type Page,
-} from "@playwright/test";
-import { createWorkspace, test as base, expect } from "./fixtures";
-
-const repoRoot = path.join(__dirname, "../..");
+  createWorkspace,
+  killApp,
+  launchApp,
+  test as base,
+  expect,
+} from "./fixtures";
 
 /**
  * Overrides the base `app` fixture to seed a projects.json before launch.
@@ -67,47 +67,9 @@ const test = base.extend<{ app: ElectronApplication; window: Page }>({
       );
     }
 
-    const app = await _electron.launch({
-      args: [path.join(repoRoot, "dist-electron/main.js")],
-      env: { ...process.env, HOME: tempHome },
-      cwd: repoRoot,
-    });
-
+    const app = await launchApp(tempHome);
     await use(app);
-
-    // Same shutdown dance as the base fixture — terminal-host child keeps the
-    // stderr pipe open so app.close() hangs. Force stdio close + SIGKILL the
-    // process group.
-    const electronProcess = app.process();
-    const pid = electronProcess.pid;
-    const closed = new Promise<void>((resolve) => {
-      if (
-        electronProcess.exitCode !== null ||
-        electronProcess.signalCode !== null
-      ) {
-        setImmediate(resolve);
-      } else {
-        electronProcess.once("close", () => setImmediate(resolve));
-      }
-    });
-    try {
-      electronProcess.stdout?.destroy();
-    } catch {
-      /* ignore */
-    }
-    try {
-      electronProcess.stderr?.destroy();
-    } catch {
-      /* ignore */
-    }
-    if (pid) {
-      try {
-        process.kill(-pid, "SIGKILL");
-      } catch {
-        /* ignore */
-      }
-    }
-    await closed;
+    await killApp(app);
   },
 });
